+++ app/Repositories/DatabasesRepository.php
@@ -2,7 +2,23 @@
 
 namespace App\Repositories;
 
-class DatabasesRepository
-{
+use Illuminate\Database\Eloquent\Model;
 
+class DatabasesRepository extends Model
+{
+    /**
+     * The table name.
+     * @var string
+     */
+    protected $table = "databases";
+
+    /**
+     * Return all data.
+     *
+     * @return mixed
+     */
+    public function index()
+    {
+        return $this->get();
+    }
 }